corrige migration: remove doc

+++ db/migrate/20210314235655_remove_column_to_academic_works.rb
@@ -1,6 +1,8 @@
 class RemoveColumnToAcademicWorks < ActiveRecord::Migration[6.1]
   def up
-    remove_column :academic_works, :doc
+    if ActiveRecord::Base.connection.column_exists?(:academic_works, :doc)
+      remove_column :academic_works, :doc
+    end
   end
 
   def down
